requirements.txt
Django==2.0
djangorestframework==3.8.2
gunicorn==19.7.1
pytz==2018.4
whitenoise==3.3.1
